+++ packages/stryker-api/core.ts
@@ -1,7 +1,8 @@
-export {default as StrykerOptions} from './src/core/StrykerOptions';
-export {default as Factory} from './src/core/Factory';
-export {default as InputFile} from './src/core/InputFile';
-export {default as Position} from './src/core/Position';
-export {default as Location} from './src/core/Location';
-export {default as Range} from './src/core/Range';
-export {default as InputFileDescriptor} from './src/core/InputFileDescriptor';
+export { default as StrykerOptions } from './src/core/StrykerOptions';
+export { default as Factory } from './src/core/Factory';
+export { default as InputFile } from './src/core/InputFile';
+export { default as Position } from './src/core/Position';
+export { default as Location } from './src/core/Location';
+export { default as Range } from './src/core/Range';
+export { default as InputFileDescriptor } from './src/core/InputFileDescriptor';
+export { default as MutationScoreThresholds } from './src/core/MutationScoreThresholds';