+++ packages/stryker-api/test/integration/install-module/install-module.ts
@@ -1,48 +1,40 @@
-import Executor from './Executor';
-import {expect} from 'chai';
+import { expect } from 'chai';
+import { exec } from 'mz/child_process';
+import * as path from 'path';
 
 describe('we have a module using stryker', function () {
 
   this.timeout(100000);
 
-  describe('after installing Stryker', () => {
-    let executor: Executor;
+  const modulePath = path.resolve(__dirname, '../../../testResources/module');
 
-    before((done) => {
-      executor = new Executor('../../../testResources/module');
-      executor.exec('npm install', {}, (errors) => done(errors));
+  const execInModule = (command: string) => {
+    console.log(`Exec '${command}' in ${modulePath}`);
+    return exec(command, { cwd: modulePath });
+  };
+
+  describe('after installing Stryker', () => {
+
+    before(() => {
+      return execInModule('npm install')
+        .then(() => execInModule('npm run tsc'));
     });
 
     describe('when typescript is compiled', () => {
 
-      let arrangeActAndAssertModule = (moduleToRun: string, partsToBeAsserted: string[]) => {
-        let stdOut: string;
-
-        describe(`and i use the "${moduleToRun}" module`, () => {
-          before((done) => {
-            executor.exec(`npm run use:${moduleToRun}`, {}, (errors, out) => {
-              stdOut = out;
-              done(errors);
-            });
-          });
-
-          it(`the output should contain ${partsToBeAsserted}`, () => {
-            partsToBeAsserted.forEach(part => expect(stdOut).to.contain(part));
+      const arrangeActAndAssertModule = (moduleToRun: string, partsToBeAsserted: string[]) => {
+        it(`should output "${partsToBeAsserted}" when using the "${moduleToRun}" module`, () => {
+          return execInModule(`npm run use:${moduleToRun}`).then(([stdout]) => {
+            partsToBeAsserted.forEach(part => expect(stdout).to.contain(part));
           });
         });
       };
-
-      before((done) => {
-        executor.exec('npm run tsc', {}, (errors) => done(errors));
-      });
-
       arrangeActAndAssertModule('core', ['files', 'some', 'file', 'pattern']);
       arrangeActAndAssertModule('config', ['plugins: [ \'stryker-*\' ]', 'port: 9234']);
       arrangeActAndAssertModule('test_framework', ['framework-1']);
       arrangeActAndAssertModule('mutant', ['nodeID: 3', 'type: \'Literal\'']);
       arrangeActAndAssertModule('report', ['empty', 'all', 'status: 3', 'originalLines: \'string\'', 'Mutant status error: Error']);
       arrangeActAndAssertModule('test_runner', ['MyTestRunner']);
-
     });
   });
 });