+++ packages/stryker-html-reporter/test/ui/hooks.ts
@@ -1,26 +1,17 @@
-import HtmlReporter from '../../src/HtmlReporter';
 import * as path from 'path';
-import { MutantResult, SourceFile } from 'stryker-api/report';
 import { browser } from 'protractor';
+import { Config } from 'stryker-api/config';
+import HtmlReporter from '../../src/HtmlReporter';
+import EventPlayer from '../helpers/EventPlayer';
 
-
-const exampleMutations = (require('../integration/exampleMutations.json') as MutantResult[]).map(mutantResult => {
-  while (mutantResult.sourceFilePath.indexOf('/') !== -1) {
-    mutantResult.sourceFilePath = mutantResult.sourceFilePath.replace('/', path.sep);
-  }
-  return mutantResult;
-});
-const exampleSourceFiles = (require('../integration/exampleSourceFiles.json') as SourceFile[]).map(file => {
-  while (file.path.indexOf('/') !== -1) {
-    file.path = file.path.replace('/', path.sep);
-  }
-  return file;
-});
+export const baseDir = path.join(__dirname, '../../reports/mutation/uiTest');
 
 before(() => {
-  let reporter = new HtmlReporter({ baseDir: 'sampleProject' });
-  reporter.onAllSourceFilesRead(exampleSourceFiles);
-  reporter.onAllMutantsTested(exampleMutations);
   browser.ignoreSynchronization = true;
-  return reporter.wrapUp();
+  const config = new Config();
+  config.set({ htmlReporter: { baseDir } });
+  const reporter = new HtmlReporter(config);
+  return new EventPlayer('testResources/mathEvents')
+    .replay(reporter)
+    .then(() => reporter.wrapUp());
 });