+++ packages/stryker/test/helpers/producers.ts
@@ -1,5 +1,16 @@
+import { Config } from 'stryker-api/config';
 import * as sinon from 'sinon';
+import { TestFramework } from 'stryker-api/test_framework';
 import { MutantStatus, MatchedMutant, MutantResult, Reporter, ScoreResult } from 'stryker-api/report';
+import { MutationScoreThresholds } from 'stryker-api/core';
+
+export type Mock<T> = {
+  [P in keyof T]: sinon.SinonStub;
+};
+
+export function mock<T>(constructorFn: { new(...args: any[]): T; }): Mock<T> {
+  return sinon.createStubInstance(constructorFn) as Mock<T>;
+}
 
 export function mutantResult(overrides: Partial<MutantResult>): MutantResult {
   const defaults: MutantResult = {
@@ -16,10 +27,21 @@ export function mutantResult(overrides: Partial<MutantResult>): MutantResult {
   return Object.assign(defaults, overrides);
 }
 
+export function testFramework(overrides?: Partial<TestFramework>): TestFramework {
+  const defaults: TestFramework = {
+    beforeEach(codeFragment: string) { return `beforeEach(){ ${codeFragment}}`; },
+    afterEach(codeFragment: string) { return `afterEach(){ ${codeFragment}}`; },
+    filter(ids: number[]) { return `filter: ${ids}`; }
+  };
+  return Object.assign(defaults, overrides);
+}
+
 export function scoreResult(score: Partial<ScoreResult>): ScoreResult {
   const defaults: ScoreResult = {
     name: 'name',
+    path: 'path',
     childResults: [],
+    representsFile: true,
     killed: 0,
     timedOut: 0,
     survived: 0,
@@ -35,6 +57,20 @@ export function scoreResult(score: Partial<ScoreResult>): ScoreResult {
   return Object.assign(defaults, score);
 }
 
+export function mutationScoreThresholds(overrides?: Partial<MutationScoreThresholds>) {
+  const defaults: MutationScoreThresholds = {
+    high: 80,
+    low: 60,
+    break: null
+  };
+  return Object.assign(defaults, overrides);
+}
+
+export function config(overrides?: Partial<Config>) {
+  const defaults: Config = new Config();
+  return Object.assign(defaults, overrides);
+}
+
 export const ALL_REPORTER_EVENTS: Array<keyof Reporter> =
   ['onSourceFileRead', 'onAllSourceFilesRead', 'onAllMutantsMatchedWithTests', 'onMutantTested', 'onAllMutantsTested', 'onScoreCalculated', 'wrapUp'];
 
@@ -45,7 +81,7 @@ export function reporterStub() {
     onAllMutantsTested: sinon.stub(),
     onAllSourceFilesRead: sinon.stub(),
     onMutantTested: sinon.stub(),
-    onScoreCalculated: sinon.stub(), 
+    onScoreCalculated: sinon.stub(),
     wrapUp: sinon.stub()
   };
 }